Workaround for texinfo versions >5.0
makeinfo 5.0+ is not compatible with this project and will break the build before completing installation. This patch is a workaround allowing the binaries to be installed as expected under CYGWIN_NT-6.1 and Ubuntu 14.04 which both uses makeinfo 5.2 from the texinfo package. However, there is still a install problem with libgcc.a which is not fixed by this patch